remove some tests.
