+++ pkg/sink/codec/avro/avro_test.go
@@ -15,13 +15,13 @@ package avro
 
 import (
 	"context"
-	"math/big"
+	"math/rand"
 	"testing"
+	"time"
 
 	"github.com/linkedin/goavro/v2"
+	timodel "github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
-	"github.com/pingcap/tidb/pkg/types"
-	"github.com/pingcap/tidb/pkg/util/rowcodec"
 	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/pkg/config"
 	"github.com/pingcap/tiflow/pkg/sink/codec/common"
@@ -30,582 +30,118 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-type avroTestColumnTuple struct {
-	col            model.Column
-	colInfo        rowcodec.ColInfo
-	expectedSchema interface{}
-	expectedData   interface{}
-	expectedType   string
-}
-
-var avroTestColumns = []*avroTestColumnTuple{
-	{
-		model.Column{Name: "tiny", Value: int64(1), Type: mysql.TypeTiny},
-		rowcodec.ColInfo{
-			ID:            1,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeTiny),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{Name: "short", Value: int64(1), Type: mysql.TypeShort},
-		rowcodec.ColInfo{
-			ID:            2,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeShort),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{Name: "int24", Value: int64(1), Type: mysql.TypeInt24},
-		rowcodec.ColInfo{
-			ID:            3,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeInt24),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{Name: "long", Value: int64(1), Type: mysql.TypeLong},
-		rowcodec.ColInfo{
-			ID:            4,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeLong),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{Name: "longlong", Value: int64(1), Type: mysql.TypeLonglong},
-		rowcodec.ColInfo{
-			ID:            5,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeLonglong),
-		},
-		avroSchema{Type: "long", Parameters: map[string]string{"tidb_type": "BIGINT"}},
-		int64(1), "long",
-	},
-	{
-		model.Column{
-			Name:  "tinyunsigned",
-			Value: uint64(1),
-			Type:  mysql.TypeTiny,
-			Flag:  model.UnsignedFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            6,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetUnsigned(types.NewFieldType(mysql.TypeTiny)),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT UNSIGNED"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{
-			Name:  "shortunsigned",
-			Value: uint64(1),
-			Type:  mysql.TypeShort,
-			Flag:  model.UnsignedFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            7,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetUnsigned(types.NewFieldType(mysql.TypeShort)),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT UNSIGNED"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{
-			Name:  "int24unsigned",
-			Value: uint64(1),
-			Type:  mysql.TypeInt24,
-			Flag:  model.UnsignedFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            8,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetUnsigned(types.NewFieldType(mysql.TypeInt24)),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "INT UNSIGNED"}},
-		int32(1), "int",
-	},
-	{
-		model.Column{
-			Name:  "longunsigned",
-			Value: uint64(1),
-			Type:  mysql.TypeLong,
-			Flag:  model.UnsignedFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            9,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetUnsigned(types.NewFieldType(mysql.TypeLong)),
-		},
-		avroSchema{Type: "long", Parameters: map[string]string{"tidb_type": "INT UNSIGNED"}},
-		int64(1), "long",
-	},
-	{
-		model.Column{
-			Name:  "longlongunsigned",
-			Value: uint64(1),
-			Type:  mysql.TypeLonglong,
-			Flag:  model.UnsignedFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            10,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetUnsigned(types.NewFieldType(mysql.TypeLonglong)),
-		},
-		avroSchema{Type: "long", Parameters: map[string]string{"tidb_type": "BIGINT UNSIGNED"}},
-		int64(1), "long",
-	},
-	{
-		model.Column{Name: "float", Value: float32(3.14), Type: mysql.TypeFloat},
-		rowcodec.ColInfo{
-			ID:            11,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeFloat),
-		},
-		avroSchema{Type: "float", Parameters: map[string]string{"tidb_type": "FLOAT"}},
-		float32(3.14), "float",
-	},
-	{
-		model.Column{Name: "double", Value: float64(3.14), Type: mysql.TypeDouble},
-		rowcodec.ColInfo{
-			ID:            12,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeDouble),
-		},
-		avroSchema{Type: "double", Parameters: map[string]string{"tidb_type": "DOUBLE"}},
-		float64(3.14), "double",
-	},
-	{
-		model.Column{Name: "bit", Value: uint64(683), Type: mysql.TypeBit},
-		rowcodec.ColInfo{
-			ID:            13,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeBit),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BIT", "length": "1"}},
-		[]byte("\x02\xab"), "bytes",
-	},
-	{
-		model.Column{Name: "decimal", Value: "129012.1230000", Type: mysql.TypeNewDecimal},
-		rowcodec.ColInfo{
-			ID:            14,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeNewDecimal),
-		},
-		avroLogicalTypeSchema{
-			avroSchema: avroSchema{
-				Type:       "bytes",
-				Parameters: map[string]string{"tidb_type": "DECIMAL"},
-			},
-			LogicalType: "decimal",
-			Precision:   10,
-			Scale:       0,
-		},
-		big.NewRat(129012123, 1000), "bytes.decimal",
-	},
-	{
-		model.Column{Name: "tinytext", Value: []byte("hello world"), Type: mysql.TypeTinyBlob},
-		rowcodec.ColInfo{
-			ID:            15,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeBlob),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{Name: "mediumtext", Value: []byte("hello world"), Type: mysql.TypeMediumBlob},
-		rowcodec.ColInfo{
-			ID:            16,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeMediumBlob),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{Name: "text", Value: []byte("hello world"), Type: mysql.TypeBlob},
-		rowcodec.ColInfo{
-			ID:            17,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeBlob),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{Name: "longtext", Value: []byte("hello world"), Type: mysql.TypeLongBlob},
-		rowcodec.ColInfo{
-			ID:            18,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeLongBlob),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{Name: "varchar", Value: []byte("hello world"), Type: mysql.TypeVarchar},
-		rowcodec.ColInfo{
-			ID:            19,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeVarchar),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{Name: "varstring", Value: []byte("hello world"), Type: mysql.TypeVarString},
-		rowcodec.ColInfo{
-			ID:            20,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeVarString),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{Name: "string", Value: []byte("hello world"), Type: mysql.TypeString},
-		rowcodec.ColInfo{
-			ID:            21,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeString),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TEXT"}},
-		"hello world", "string",
-	},
-	{
-		model.Column{
-			Name:  "tinyblob",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeTinyBlob,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            22,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeTinyBlob)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{
-			Name:  "mediumblob",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeMediumBlob,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            23,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeMediumBlob)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{
-			Name:  "blob",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeBlob,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            24,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeBlob)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{
-			Name:  "longblob",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeLongBlob,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            25,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeLongBlob)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{
-			Name:  "varbinary",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeVarchar,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            26,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeVarchar)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{
-			Name:  "varbinary1",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeVarString,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            27,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeVarString)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{
-			Name:  "binary",
-			Value: []byte("hello world"),
-			Type:  mysql.TypeString,
-			Flag:  model.BinaryFlag,
-		},
-		rowcodec.ColInfo{
-			ID:            28,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetBinChsClnFlag(types.NewFieldType(mysql.TypeString)),
-		},
-		avroSchema{Type: "bytes", Parameters: map[string]string{"tidb_type": "BLOB"}},
-		[]byte("hello world"), "bytes",
-	},
-	{
-		model.Column{Name: "enum", Value: uint64(1), Type: mysql.TypeEnum},
-		rowcodec.ColInfo{
-			ID:            29,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetElems(types.NewFieldType(mysql.TypeEnum), []string{"a", "b"}),
-		},
-		avroSchema{
-			Type:       "string",
-			Parameters: map[string]string{"tidb_type": "ENUM", "allowed": "a,b"},
-		},
-		"a", "string",
-	},
-	{
-		model.Column{Name: "set", Value: uint64(1), Type: mysql.TypeSet},
-		rowcodec.ColInfo{
-			ID:            30,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            utils.SetElems(types.NewFieldType(mysql.TypeSet), []string{"a", "b"}),
-		},
-		avroSchema{
-			Type:       "string",
-			Parameters: map[string]string{"tidb_type": "SET", "allowed": "a,b"},
-		},
-		"a", "string",
-	},
-	{
-		model.Column{Name: "json", Value: `{"key": "value"}`, Type: mysql.TypeJSON},
-		rowcodec.ColInfo{
-			ID:            31,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeJSON),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "JSON"}},
-		`{"key": "value"}`, "string",
-	},
-	{
-		model.Column{Name: "date", Value: "2000-01-01", Type: mysql.TypeDate},
-		rowcodec.ColInfo{
-			ID:            32,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeDate),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "DATE"}},
-		"2000-01-01", "string",
-	},
-	{
-		model.Column{Name: "datetime", Value: "2015-12-20 23:58:58", Type: mysql.TypeDatetime},
-		rowcodec.ColInfo{
-			ID:            33,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeDatetime),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "DATETIME"}},
-		"2015-12-20 23:58:58", "string",
-	},
-	{
-		model.Column{Name: "timestamp", Value: "1973-12-30 15:30:00", Type: mysql.TypeTimestamp},
-		rowcodec.ColInfo{
-			ID:            34,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeTimestamp),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TIMESTAMP"}},
-		"1973-12-30 15:30:00", "string",
-	},
-	{
-		model.Column{Name: "time", Value: "23:59:59", Type: mysql.TypeDuration},
-		rowcodec.ColInfo{
-			ID:            35,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeDuration),
-		},
-		avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "TIME"}},
-		"23:59:59", "string",
-	},
-	{
-		model.Column{Name: "year", Value: int64(1970), Type: mysql.TypeYear},
-		rowcodec.ColInfo{
-			ID:            36,
-			IsPKHandle:    false,
-			VirtualGenCol: false,
-			Ft:            types.NewFieldType(mysql.TypeYear),
-		},
-		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "YEAR"}},
-		int32(1970), "int",
-	},
-}
-
-func TestColumnToAvroSchema(t *testing.T) {
-	for _, v := range avroTestColumns {
-		encoder := NewAvroEncoder("namespace", nil, &common.Config{
-			AvroDecimalHandlingMode:        "precise",
-			AvroBigintUnsignedHandlingMode: "long",
-		})
-		schema, err := encoder.(*BatchEncoder).columnToAvroSchema(&v.col, v.colInfo.Ft)
-		require.NoError(t, err)
-		require.Equal(t, v.expectedSchema, schema)
-		if v.col.Name == "decimal" {
-			encoder := NewAvroEncoder("namespace", nil, &common.Config{
-				AvroDecimalHandlingMode:        "string",
-				AvroBigintUnsignedHandlingMode: "long",
-			})
-			schema, err := encoder.(*BatchEncoder).columnToAvroSchema(&v.col, v.colInfo.Ft)
-			require.NoError(t, err)
-			require.Equal(
-				t,
-				avroSchema{Type: "string", Parameters: map[string]string{"tidb_type": "DECIMAL"}},
-				schema,
-			)
-		}
-		if v.col.Name == "longlongunsigned" {
-			encoder := NewAvroEncoder("namespace", nil, &common.Config{
-				AvroDecimalHandlingMode:        "precise",
-				AvroBigintUnsignedHandlingMode: "string",
-			})
-			schema, err := encoder.(*BatchEncoder).columnToAvroSchema(&v.col, v.colInfo.Ft)
-			require.NoError(t, err)
-			require.Equal(
-				t,
-				avroSchema{
-					Type:       "string",
-					Parameters: map[string]string{"tidb_type": "BIGINT UNSIGNED"},
-				},
-				schema,
-			)
-		}
-	}
-}
-
-func TestColumnToAvroData(t *testing.T) {
-	t.Parallel()
-
-	for _, v := range avroTestColumns {
-		encoder := NewAvroEncoder("namespace", nil, &common.Config{
-			AvroDecimalHandlingMode:        "precise",
-			AvroBigintUnsignedHandlingMode: "long",
-		})
-		data, str, err := encoder.(*BatchEncoder).columnToAvroData(&v.col, v.colInfo.Ft)
-		require.NoError(t, err)
-		require.Equal(t, v.expectedData, data)
-		require.Equal(t, v.expectedType, str)
-		if v.col.Name == "decimal" {
-			encoder := NewAvroEncoder("namespace", nil, &common.Config{
-				AvroDecimalHandlingMode:        "string",
-				AvroBigintUnsignedHandlingMode: "long",
-			})
-			data, str, err := encoder.(*BatchEncoder).columnToAvroData(&v.col, v.colInfo.Ft)
-			require.NoError(t, err)
-			require.Equal(t, "129012.1230000", data)
-			require.Equal(t, "string", str)
-		}
-		if v.col.Name == "longlongunsigned" {
-			encoder := NewAvroEncoder("namespace", nil, &common.Config{
-				AvroDecimalHandlingMode:        "precise",
-				AvroBigintUnsignedHandlingMode: "string",
-			})
-			data, str, err := encoder.(*BatchEncoder).columnToAvroData(&v.col, v.colInfo.Ft)
-			require.NoError(t, err)
-			require.Equal(t, "1", data)
-			require.Equal(t, "string", str)
-		}
-	}
-}
-
-func TestRowToAvroData(t *testing.T) {
-	_, event, _, _ := utils.NewLargeEvent4Test(t, config.GetDefaultReplicaConfig())
-	columns := event.GetColumns()
-	colInfos := event.TableInfo.GetColInfosForRowChangedEvent()
-	input := &avroEncodeInput{
-		columns,
-		colInfos,
-	}
-
+func TestDMLEventE2E(t *testing.T) {
 	codecConfig := common.NewConfig(config.ProtocolAvro)
+	codecConfig.EnableTiDBExtension = true
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	_, event, _, _ := utils.NewLargeEvent4Test(t, config.GetDefaultReplicaConfig())
+	colInfos := event.TableInfo.GetColInfosForRowChangedEvent()
+
+	ra
+	nd.New(rand.NewSource(time.Now().Unix())).Shuffle(len(event.Columns), func(i, j int) {
+		event.Columns[i], event.Columns[j] = event.Columns[j], event.Columns[i]
+		colInfos[i], colInfos[j] = colInfos[j], colInfos[i]
+	})
+
+	for _, decimalHandling := range []string{"precise", "string"} {
+		for _, unsignedBigintHandling := range []string{"long", "string"} {
+			codecConfig.AvroDecimalHandlingMode = decimalHandling
+			codecConfig.AvroBigintUnsignedHandlingMode = unsignedBigintHandling
+
+			encoder, err := SetupEncoderAndSchemaRegistry4Testing(ctx, codecConfig)
+			require.NoError(t, err)
+			require.NotNil(t, encoder)
+
+			topic := "avro-test-topic"
+			err = encoder.AppendRowChangedEvent(ctx, topic, event, func() {})
+			require.NoError(t, err)
+
+			messages := encoder.Build()
+			require.Len(t, messages, 1)
+			message := messages[0]
+
+			schemaM, err := NewConfluentSchemaManager(ctx, "http://127.0.0.1:8081", nil)
+			require.NoError(t, err)
+
+			decoder := NewDecoder(codecConfig, schemaM, topic)
+			err = decoder.AddKeyValue(message.Key, message.Value)
+			require.NoError(t, err)
+
+			messageType, exist, err := decoder.HasNext()
+			require.NoError(t, err)
+			require.True(t, exist)
+			require.Equal(t, model.MessageTypeRow, messageType)
+
+			decodedEvent, err := decoder.NextRowChangedEvent()
+			require.NoError(t, err)
+			require.NotNil(t, decodedEvent)
+
+			TeardownEncoderAndSchemaRegistry4Testing()
+		}
+	}
+}
+
+func TestDDLEventE2E(t *testing.T) {
+	codecConfig := common.NewConfig(config.ProtocolAvro)
+	codecConfig.EnableTiDBExtension = true
+	codecConfig.AvroEnableWatermark = true
+
 	encoder := NewAvroEncoder(model.DefaultNamespace, nil, codecConfig)
 
-	data, err := encoder.(*BatchEncoder).columns2AvroData(input)
+	ddl, _, _, _ := utils.NewLargeEvent4Test(t, config.GetDefaultReplicaConfig())
+	message, err := encoder.EncodeDDLEvent(ddl)
+	require.NoError(t, err)
+	require.NotNil(t, message)
+
+	topic := "test-topic"
+	decoder := NewDecoder(codecConfig, nil, topic)
+	err = decoder.AddKeyValue(message.Key, message.Value)
 	require.NoError(t, err)
 
-	for _, col := range input.columns {
-		_, exists := data[col.Name]
-		require.True(t, exists)
-	}
+	messageType, exist, err := decoder.HasNext()
+	require.NoError(t, err)
+	require.True(t, exist)
+	require.Equal(t, model.MessageTypeDDL, messageType)
+
+	decodedEvent, err := decoder.NextDDLEvent()
+	require.NoError(t, err)
+	require.NotNil(t, decodedEvent)
+	require.Equal(t, ddl.CommitTs, decodedEvent.CommitTs)
+	require.Equal(t, timodel.ActionCreateTable, decodedEvent.Type)
+	require.NotEmpty(t, decodedEvent.Query)
+	require.NotEmpty(t, decodedEvent.TableInfo.TableName.Schema)
+	require.NotEmpty(t, decodedEvent.TableInfo.TableName.Table)
+}
+
+func TestResolvedE2E(t *testing.T) {
+	t.Parallel()
+
+	codecConfig := common.NewConfig(config.ProtocolAvro)
+	codecConfig.EnableTiDBExtension = true
+	codecConfig.AvroEnableWatermark = true
+
+	encoder := NewAvroEncoder(model.DefaultNamespace, nil, codecConfig)
+
+	resolvedTs := uint64(1591943372224)
+	message, err := encoder.EncodeCheckpointEvent(resolvedTs)
+	require.NoError(t, err)
+	require.NotNil(t, message)
+
+	topic := "test-topic"
+	decoder := NewDecoder(codecConfig, nil, topic)
+	err = decoder.AddKeyValue(message.Key, message.Value)
+	require.NoError(t, err)
+
+	messageType, exist, err := decoder.HasNext()
+	require.NoError(t, err)
+	require.True(t, exist)
+	require.Equal(t, model.MessageTypeResolved, messageType)
+
+	obtained, err := decoder.NextResolvedEvent()
+	require.NoError(t, err)
+	require.Equal(t, resolvedTs, obtained)
 }
 
 func TestAvroEncode4EnableChecksum(t *testing.T) {
